Create OpenCV.vpr
Created config that will be used for exporting the image to the OpenCV algorithm. The config disables LimeLight's object detection but the threshold & exposure/black light offset are perfect. It can clearly see the target @ 5 or 40 feet away. It does pick up othe objects, but with OpenCV they won't be a problem.